move auth stuff to firebase-queries.js
