+++ src/firebase-config.js
@@ -1,9 +1,6 @@
 import { initializeApp } from "firebase/app"
 import { getFirestore } from "firebase/firestore"
-import { getAuth, onAuthStateChanged } from "firebase/auth"
-import { store } from "./store/store"
-
-import { setLoading, setUser } from "./store/actions"
+import { getAuth } from "firebase/auth"
 
 const firebaseConfig = {
   apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
@@ -19,17 +16,3 @@ const app = initializeApp(firebaseConfig)
 
 export const db = getFirestore(app)
 export const auth = getAuth(app)
-
-export const authStatus = async () => {
-  store.dispatch(setLoading(true))
-
-  onAuthStateChanged(auth, (user) => {
-    if (user !== null) {
-      store.dispatch(setUser(user))
-    } else {
-      store.dispatch(setUser({}))
-    }
-
-    store.dispatch(setLoading(false))
-  })
-}
